+++ CHANGELOG.md
@@ -6,8 +6,12 @@
 - Support offset for each input/output buffer in Tensile
 - support support ldc != ldd for all GEMM kernel
 
+### Optimizations
+- Refactor ConvolutionVsContraction
+
 ### Fixed
 - Fixed MasterSolutionLibrary having duplicated hardware rows
+- channel stride is incorrect when converting conv problem into tensor contraction problem
 
 ## [Tensile 4.26.0 for ROCm 4.1.0]
 ### Added